allow group by in recursive clause of RCTEs
This is an experimental feature. Its acceptance in core Postgres was
discussed, and rejected, in [1].
It allows to mimic the behavior of early aggregates in recursive
CTEs (UNION ALL only). For this purpose, it only works for trivial aggregates for which
the state type is the same as the final type.

[1] https://www.postgresql.org/message-id/flat/CAN_p2QgUckBduWmjid8rB4FVUvK9q5UmB%3Din0f-d%3DhTp_0JXmw%40mail.gmail.com