Refine error messages regarding recursive CTE restrictions
These clauses are attached to the UNION clause in:

WITH RECURSIVE ... (... UNION ... ORDER BY ) ...

In this case, it is not allowed and it will not be
implemented. However, the following case is supported:

WITH RECURSIVE ... (... UNION (... ORDER BY) ) ...

The previous message led to think that this case is not supported
neither.